参数更新

+++ hsd-account/hsd-account-staff/hsd-account-staff-services/hsd-account-staff-biz/src/main/java/com/hsd/account/staff/service/sys/SysAppService.java
@@ -60,9 +60,6 @@ public class SysAppService extends BaseService implements ISysAppService {
         try {
             if (dto == null) throw new RuntimeException("参数异常!");
             SysApp entity = copyTo(dto, SysApp.class);
-            if (sysAppDao.isNoDelFlag(entity) > 0) {
-                throw new RuntimeException("当前数据为系统保留数据，不能删除!");
-            }
             if (sysAppDao.deleteByPrimaryKey(entity) == 0) {
                 throw new RuntimeException("数据不存在!");
             }
@@ -78,6 +75,9 @@ public class SysAppService extends BaseService implements ISysAppService {
         try {
             if (dto == null) throw new RuntimeException("参数异常!");
             SysApp entity = copyTo(dto, SysApp.class);
+            if (sysAppDao.isNoDelFlag(entity) > 0) {
+                throw new RuntimeException("当前数据为系统保留数据，不能删除!");
+            }
             if (sysAppDao.deleteById(entity) == 0) {
                 throw new RuntimeException("数据不存在!");
             }
